extract/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
requests==2.22.0
six==1.15.0
typed-ast==1.4.0
urllib3==1.24.2
wrapt==1.11.2
xlrd==1.2.0
google-cloud-storage==1.35.0
jsonschema==3.2.0
